+++ core/src/main/scala/com/krux/hyperion/Schedule.scala
@@ -42,7 +42,7 @@ case class Schedule private (
 
   def every(p: Duration) = this.copy(period = p)
   def until(dt: DateTime) = this.copy(end = Option(Right(dt)))
-  def stopAfter(occurrences: Int) = this.copy(end = Option(Left(occurrences)))
+  def stopAfter(occurrences: Int) = this.copy(end = if (occurrences > 0) Option(Left(occurrences)) else None)
 
   def objects: Iterable[PipelineObject] = None