+++ core/src/main/scala/com/krux/hyperion/activity/RedshiftCopyOption.scala
@@ -1,37 +1,56 @@
 package com.krux.hyperion.activity
 
-trait RedshiftCopyOption {
-  def repr: Seq[String]
-}
+case class RedshiftCopyOption(repr: Seq[String])
 
 object RedshiftCopyOption {
 
-  def csv(quote: String) = new RedshiftCopyOption {
-    def repr = Seq("CSV", "QUOTE", quote)
-  }
+  def quoted[T](x: T): String = s"'$x'"
 
-  def csv = new RedshiftCopyOption {
-    def repr = Seq("CSV")
-  }
+  def csv(quote: String) = RedshiftCopyOption(Seq("CSV", "QUOTE", quoted(quote)))
 
-  def gzip = new RedshiftCopyOption {
-    def repr = Seq("GZIP")
-  }
+  def csv = RedshiftCopyOption(Seq("CSV"))
 
-  def delimiter(delChar: String) = new RedshiftCopyOption {
-    def repr = Seq("DELIMITER", s"'$delChar'")
-  }
+  def gzip = RedshiftCopyOption(Seq("GZIP"))
 
-  def escape = new RedshiftCopyOption {
-    def repr = Seq("ESCAPE")
-  }
+  def delimiter(delChar: String) = RedshiftCopyOption(Seq("DELIMITER", quoted(delChar)))
 
-  def nullAs(nullStr: String) = new RedshiftCopyOption {
-    def repr = Seq("NULL", s"'$nullStr'")
-  }
+  def escape = RedshiftCopyOption(Seq("ESCAPE"))
 
-  def maxError(errorCount: Int) = new RedshiftCopyOption {
-    def repr = Seq("MAXERROR", errorCount.toString)
-  }
+  def nullAs(nullStr: String) = RedshiftCopyOption(Seq("NULL", quoted(nullStr)))
 
+  def maxError(errorCount: Int) = RedshiftCopyOption(Seq("MAXERROR", errorCount.toString))
+
+  def acceptInvChars = RedshiftCopyOption(Seq("ACCEPTINVCHARS"))
+
+  def acceptInvChars(replacementChar: Char) =
+    RedshiftCopyOption(Seq("ACCEPTINVCHARS", quoted(replacementChar)))
+
+  def acceptAnyDate = RedshiftCopyOption(Seq("ACCEPTANYDATE"))
+
+  def blanksAsNull = RedshiftCopyOption(Seq("BLANKSASNULL"))
+
+  def dateFormat = RedshiftCopyOption(Seq("DATEFORMAT"))
+
+  def dateFormat(dateFormatString: String) =
+    RedshiftCopyOption(Seq("DATEFORMAT", quoted(dateFormatString)))
+
+  def encoding(fileEncoding: String) = RedshiftCopyOption(Seq("ENCODING", fileEncoding))
+
+  def explicitIds = RedshiftCopyOption(Seq("EXPLICIT_IDS"))
+
+  def fillRecord = RedshiftCopyOption(Seq("FILLRECORD"))
+
+  def ignoreBlankLines = RedshiftCopyOption(Seq("IGNOREBLANKLINES"))
+
+  def ignoreHeader(numberRows: Long) = RedshiftCopyOption(Seq("IGNOREHEADER", numberRows.toString))
+
+  def removeQuotes = RedshiftCopyOption(Seq("REMOVEQUOTES"))
+
+  def roundec = RedshiftCopyOption(Seq("ROUNDEC"))
+
+  def timeFormat(format: String) = RedshiftCopyOption(Seq("TIMEFORMAT", quoted(format)))
+
+  def trimBlanks = RedshiftCopyOption(Seq("TRIMBLANKS"))
+
+  def truncateColumns = RedshiftCopyOption(Seq("TRUNCATECOLUMNS"))
 }